backend/requirements.txt
Django==3.2.3
django-cors-headers==4.3.1
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.2.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1 
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.21.0
Pillow
whitenoise==3.3.1